Test: Remove check for unused coeffIdx in ipfilter tests
The luma (8-tap) and chroma (4-tap) interpolation primitives are only
called with non-zero values of coeffIdx. This patch alters the tests for
these primitives so that they do not check with a coeffIdx of zero, in
order to allow this unused path to be omitted from SIMD optimised
primitives without causing test failures.