Avoid pointer copies in ReferencePlanes::{lowresMC,lowresQPelCost}
These copies are reasonably hot in perf traces and there is no need to
copy the pointers here. Just set a pointer to the array to read from and
use that directly.

Benchmarks show this gives a nice improvement in FPS reported for an
encoding using --preset=medium:

Cortex-A715: +0.2%
  Cortex-X3: +0.7%
Neoverse V2: +0.9%

Change-Id: I33ba6854e7fbe6ecb1778b8201030a717f60f34e