AArch64: Optimise HBD interp_horiz_pp_neon
Optimise the HBD 4-tap and 8-tap Neon implementations of
interp_horiz_pp_neon and extend these functions to support all CHROMA
and LUMA block sizes respectively.

The new 4-tap filter implementation is up to 37% faster when
coeffIdx==4 and up to 20% faster for the other filter values compared
to the existing Neon implementation.

The new 8-tap filter implementation is up to 42% faster when
coeffIdx==1, 51% when it is 2, and 44% when it is 3; compared to the
existing Neon implementation.