Avoid aliasing HBD SSE_PP functions for AArch64 platforms
Avoid using AArch64 sse_ss implementation for sse_pp in high
bit-depth builds. Even though the parameter for both functions is
16-bit, sse_ss uses signed parameters as opposed to sse_pp which uses
unsigned ones. Because of this difference, in subsequent patches
different AArch64 implementations will be added for these functions.
The sign bit matters in the context of what the maximum value of SSE
can be and the number of accumulator registers.

Change-Id: I21ac7249f8968eedca2da39f7cd11b6489bac6cc