AArch64: Enable existing SSE_SS SVE impl for SBD
The existing HBD SSE_SS SVE implementation is suitable for SBD as
well, so enable it for the SBD build.

Delete the existing SBD SSE_SS SVE2 implementation in order to have
the SVE implementation as default for SVE2 supported platforms. The
SVE implementation is up to 55% faster than the SVE2 implementation.

Change-Id: Ib7f5e731b2007ebd6e967cfb1a5ffbeb845dac22