AArch64: Enable existing SSD_S SVE impl for SBD
The existing HBD SSD_S SVE implementation is suitable for SBD as
well, so enable it for the SBD build.

Delete the existing SBD SSD_S SVE2 implementation in order to have
the SVE implementation as default for SVE2 supported platforms. The
SVE implementation is up to 55% faster than the SVE2 implementation.

Change-Id: I0fc349d1c875f74743a956869725ceecc5825ee1