testbench.cpp: Add --cpuid list to list valid SIMD architectures
Rather than needing to guess what architecture features are available to
be tested, simply allow the user to query the testbench to discover the
available features. Since the list of architectures is now a global
variable, adjust the naming to camelCase instead of snake_case to match
the rest of the file.

Also add the relevant explanation line to --help text.

Also take this opportunity to reorder SVE/SVE2 in the Arm architecture
list such that they occur below the group of Neon extensions.