asm-primitives.cpp: Delete dequant_scaling SVE2 implementation
The implementation makes use of Neon instructions which only operate on
the low 128-bits of an SVE register, meaning that for longer vectors the
function gives incorrect output. We could maintain the kernel and only
use it if the vector length is 128-bits, however it appears to give no
benefit in this case so simply remove it instead.