mc-a-sve2.S: Fix addAvg_{16,32}xh_sve2 for longer SVE vectors
The 256-bit SVE vector length loop for 16xh already only processes one
full vector per iteration, so a further specialization for even longer
vectors cannot be useful. In particular the implementation here for
vector lengths of at least 512 bits appears to be incorrect and
unreachable, so simply delete it and use the 256-bit implmentation
instead.

The existing code for 32xh flips the add and sqrshrnb instructions
leading to an incorrect result, so reorder them to fix.

Co-authored-by: Hari Limaye <hari.limaye@arm.com>