AArch64: Optimise HBD interp_horiz_pp_neon for 10-bit
Optimise the HBD 4-tap and 8-tap Neon implementations of
interp_horiz_pp_neon when X265_DEPTH is 10.

The 4-tap filter implementation is up to 28% faster when
coeffIdx==4 and up to 8% faster for the other filter values compared
to the existing Neon implementation.

The 8-tap filter implementation is up to 17% faster when
coeffIdx is 1 or 3 compared to the existing Neon implementation.

The existing high bitdepth Neon implementation is retained for use
with 12-bit input.