AArch64: Optimise HBD interp_horiz_ps_neon
Optimise the HBD 4-tap and 8-tap Neon implementations of
interp_horiz_ps_neon and extend these functions to support all CHROMA
and LUMA block sizes respectively.

The new 4-tap filter implementation is up to 34% faster when
coeffIdx==4 and up to 11% faster for the other filter values compared
to the existing Neon implementation.

The new 8-tap filter implementation is up to 34% faster when
coeffIdx==1, 48% when it is 2, and 40% when it is 3; compared to the
existing Neon implementation.