AArch64: Optimise HBD interp_vert_pp_neon
Optimise the HBD 4-tap and 8-tap Neon implementations of
interp_vert_pp_neon and extend these functions to support all CHROMA
and LUMA block sizes respectively.

For the 12-bit implementation the new 4-tap filter implementation is
up to 51% faster when coeffIdx==4 and up to 33% faster for the other
filter values compared to the existing Neon implementation. The new
8-tap filter implementation is up to 55% faster when coeffIdx==1,
64% when it is 2, and 58% when it is 3; compared to the existing Neon
implementation.

For the 10-bit implementation the new 4-tap filter implementation is
up to 25% faster when coeffIdx==4 and up to 17% faster for the other
filter values compared to the new 12-bit Neon implementation. The
8-tap filter implementation is up to 16% faster when coeffIdx is 1 or
3 compared to the new 12-bit Neon implementation.