AArch64: Optimise HBD interp_vert_ps_neon
Optimise the HBD 4-tap and 8-tap Neon implementations of
interp_vert_ps_neon and extend these functions to support all CHROMA
and LUMA block sizes respectively.

For the 12-bit implementation the new 4-tap filter implementation is
up to 54% faster when coeffIdx==4 and up to 28% faster for the other
filter values compared to the existing Neon implementation. The new
8-tap filter implementation is up to 55% faster when coeffIdx==1,
64% when it is 2, and 58% when it is 3; compared to the existing Neon
implementation.

For the 10-bit implementation the new 4-tap filter implementation is
up to 58% faster when coeffIdx==4 and up to 49% faster for the other
filter values compared to the new 12-bit Neon implementation. The
8-tap filter implementation is up to 23% faster when coeffIdx is 1 or
18% when it is 3 compared to the new 12-bit Neon implementation.