AArch64: Optimise SBD interp_vert_ss/sp
Optimise the SBD 4-tap and 8-tap filter application for
interp_vert_sp_neon and interp_vert_ss_neon and rewrite the control
flow for these functions to optimise width < 8 paths.

The new 4-tap filter implementation for interp_vert_ss_neon is up to
22% faster when coeffIdx==4 compared to the existing Neon
implementation. The new 8-tap filter implementation is up to 8%
faster when coeffIdx==1, 23% when it is 2, and 9% when it is 3;
compared to the existing Neon implementation.

The new 4-tap filter implementation for interp_vert_sp_neon is up to
8% faster when coeffIdx==4 compared to the existing Neon
implementation. The new 8-tap filter implementation is up to 5%
faster when coeffIdx==1, 25% when it is 2, and 7% when it is 3;
compared to the existing Neon implementation.