AArch64: Add SVE2_BitPerm implementation of scanPosLast
Use the BEXT instruction to pack bits based on a bitmask, this avoids
the need for the loop in the prior Arm implementations.

There was an existing function declaration of scanPosLast_sve2, however
this function was never defined and unused so simply replace it with the
new scanPosLast_sve2_bitperm declaration.

Benchmarking on a Neoverse V2 machine supporting the SVE2_BitPerm
extension, this implementation improves --preset=medium encoding speed
by ~1.3%.

Also take this opportunity to reorder the ARM_ASMS extension list in
CMakeLists.txt to be in architecture order to match elsewhere.