AArch64: Add SVE BitPerm enablement and feature detection
This patch expands the existing CMake and runtime feature detection code
to include support for the AArch64 SVE Bit Permute extension
(FEAT_SVE_BitPerm). This is an optional extension to SVE2, introduced
from Armv9.0-A.

In particular this extension provides support for the BEXT, BDEP, and
BGRP instructions which can be useful when manipulating bit masks. This
includes kernels like scanPosLast for which an optimized implementation
making use of this extension will follow in a later commit.