AArch64: Add SVE implementation of HBD interp_horiz_pp
Add SVE implementation of HBD interp_horiz_pp for LUMA filtering.
An implementation was added for block sizes with width equal to 4 for
both 10-bit and 12-bit build, but for bigger block sizes the SVE
implementation was only enabled for 12-bit build.

This implementation gives up to 9% uplift compared to the existing
Neon implementation.