AArch64: Add SBD and HBD Neon implementation of planecopy_cp
Add Neon intrinsic implementation of planecopy_cp that works for both
SBD and HBD. Remove the Neon asm for SBD as the intrinsics
implementation provides better performance.

This Neon implementation is optimised such that the shift value is fixed
as 'X265_DEPTH - 8' to reflect the behaviour of the actual call to the
function.

Relative performance compared to Neon asm [SBD]:
 Neoverse N1: 1.35x
 Neoverse N2: 0.92x
 Neoverse V1: 1.09x
 Neoverse V2: 1.28x

Relative performance compared to scalar C [HBD]:
 Neoverse N1: 2.90x
 Neoverse N2: 3.23x
 Neoverse V1: 3.46x
 Neoverse V2: 3.00x