AArch64: Optimize HBD interp_hv_pp_neon function
The existing interp_hv_pp_neon function performs horizontal
interpolation, stores the result in an intermediate buffer, and then
applies vertical interpolation on that buffer. This can be optimized
by merging the horizontal and vertical interpolation eliminating the
need for the intermediate buffer.

The optimisation is applicable to all block sizes and gives an uplift
of up to 10% for 10-bit builds and 13% for 12-bit builds.