AArch64: Simplify SBD and HBD psyCost_pp_neon
Simplify the psyCost_pp_neon implementation by removing unnecessary
computation of absolute differences against zeroBuf in the calls to the
SA8D/SATD and SAD functions. These are replaced with a combined function
where both Hadamard 4/8 and summations are calculated and subtracted.

Also remove the assembly implementation for 4x4 block size which is
slower than the simplified intrinsics.