AArch64: Optimize SBD interp_hv_pp_i8mm function
The existing interp_hv_pp_i8mm function performs horizontal
interpolation, stores the result in an intermediate buffer, and then
applies vertical interpolation on that buffer. This can be optimized
by merging the horizontal and vertical interpolation eliminating the
need for the intermediate buffer.

This optimization is applicable to all block sizes and gives a
performance uplift of up to 9%.

Co-authored by: Arpad Panyik arpad.panyik@arm.com