AArch64: Optimize SBD interp_hv_pp_dotprod function
The existing interp_hv_pp_dotprod function performs horizontal
interpolation, stores the result in an intermediate buffer, and then
applies vertical interpolation on that buffer. This can be optimized
by merging the horizontal and vertical interpolation eliminating the
need for the intermediate buffer.

This optimization is applicable to some of the block sizes with width
4 and gives a performance uplift of up to 5%.

Co-authored by: Arpad Panyik arpad.panyik@arm.com