AArch64: Fix SVE 16x16 and 32x32 DCT implementations
int16_t elements do not provide sufficient range to compute the 'EO'
data. Refactor the SVE path to use the same int32_t element approach
for computing 'EO' as in the Neon path.

This bug was found after observing differences in the encoded video
output for various speed presets and video resolutions. Unit tests
did not catch this issue.