AArch64: Specialize passes of 16x16 and 32x32 SVE DCTs
A previous patch fixed bugs in the 16x16 and 32x32 SVE paths, but
this fix brought a performance penalty. Regain some of the lost
performance in this patch by specializing the first pass of the 16x16
and 32x32 DCT paths to take advantage of the lower intermediate
precision requirements - remaining at int16_t elements when computing
'EO' such that we can use the faster SDOT approach later.