AArch64: Optimize and clean up addAvg Neon and SVE2 functions
Extend the neon intrinsics implementation to support all block sizes and
optimize it to use rounding-shift-and-accumulate instead of separate
widening, add, and shift steps. Also unroll the loops for larger block
sizes to enable the compiler to emit LDP and STP instructions.

Delete the Neon and SVE2 assembly implementations as they are 1-2x
slower than Neon intrinsics implementation.