AArch64: Fix scanPosLast_neon compilation with old binutils
When compiling using GNU Binutils version 2.38 (the default on Ubuntu
22.04), the assembler does not recognise the AArch64 BFC instruction
unless Armv8.2-A is enabled. This causes the scanPosLast_neon function
to fail to compile. This appears to be fixed with GNU Binutils 2.42 (the
default on Ubuntu 24.04).

The BFC instruction is an alias of BFM, so rewrite it to express with a
BFM instruction instead. This ensures that it compiles even on the older
GNU Binutils version.